core: fix agentx message parsing on big endian machines
When parsing agentx messages we not only need to consider the endianess
from our peer, which it provides via the PDU flag, but also our own.

Only if our and the peer's endianess differ then we need to swap bytes.

Both the provided NTOH16/32/64 functions provided by smithsnmp and the
posix compatible htons()/htonl() functions are not able to do this. The
current NTOH16/32/64 implementation only considers the foreign but not
our own endianess. htons()/htonl() on the hand would only consider our
own but not the foreign endianess.

Therefore replace the current NTOH16/32/64 and HTON16/32/64 macros with
axtoh16/32/64 htoax16/32/64 which are capable of byte swapping if the
own and the foreign endianess differ.

While at it, replace the byteswaping implemented in smithsnmp with
implementations provided by the OS.

Kudos to Alexander Sarmanow for his initial investigations and work
and pointers to the right direction.

Cc: Alexander Sarmanow <as@simonwunderlich.de>
Signed-off-by: Linus Lüssing <ll@simonwunderlich.de>