scons: use Python native endianess check to fix cross compiling
When cross compiling smithsnmpd, for instance for OpenWrt, then the
inline C code of endian_probe will be compiled for the target
architecture. And the subsequent endianess check from SConstruct will
fail with an Exec format error because the binary format matches the
target and not the host architecture:

  Checking for Endian...sh: 1: .sconf_temp/conftest_0: Exec format error

Fixing this by replacing the inline C code in Python with Python's native
endianess functions.

Signed-off-by: Linus Lüssing <ll@simonwunderlich.de>