scons: avoid overriding endianess from environment, fix cross-compile
From within SConstruct the Python sys.byteorder will only provide the
byte order of the host system and not the byte order of the target
we might cross-compile for.

As the previous approach of compiling a small C program from Python
and trying to run it from the host does not work either, the alternative
is to simply allow the build system to provide the endianess of the
target we are building for.

With this the build system can now provide "-DLITTLE_ENDIAN" or
"-DBIG_ENDIAN" via CFLAGS. If specified, SConstruct will now refrain
from trying to determine the endianess on its own.

Signed-off-by: Linus Lüssing <ll@simonwunderlich.de>